backend/requirements.txt
flask==2.0.1
flask-pymongo==2.3.0
flask-cors==3.0.10
python-dotenv==0.19.1
pymongo==4.0.1
Werkzeug==2.0.1
gunicorn==20.1.0
dnspython==2.3.0
PyJWT==2.8.0
